Add basic json endpoint for recipes

+++ migrations/20231118081028_recipes.up.sql
@@ -21,7 +21,7 @@ CREATE TABLE recipe_ingredients (
     id INTEGER PRIMARY KEY,
     recipe_id INTEGER NOT NULL,
     ingredient_id INTEGER NOT NULL,
-    amount DECIMAL NOT NULL,
+    amount FLOAT NOT NULL,
     unit TEXT NOT NULL,
     FOREIGN KEY (recipe_id) REFERENCES recipes(id),
     FOREIGN KEY (ingredient_id) REFERENCES ingredients(id)
